Extract Tokenizer Escaping tests out to a separate file
