+++ super-csv/src/test/java/org/supercsv/io/TokenizerTest.java
@@ -29,7 +29,6 @@ import java.util.ArrayList;
 import java.util.List;
 
 import org.junit.After;
-import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
 import org.supercsv.comment.CommentMatches;
@@ -586,201 +585,6 @@ public class TokenizerTest {
 		assertEquals("  two  ", columns.get(1));
 		assertEquals("   three   ", columns.get(2));
 		assertEquals(input, tokenizer.getUntokenizedRow());
-	}
-
-	/**
-	 * Tests the readColumns() method with a quote character escaped with another quote (RFC style)
-	 */
-	@Test
-	public void testQuotedQuoteCharWithDoubleQuoteEscape() throws Exception {
-		final String input = "one,\"two\",\"field with \"\" quote char\",four";
-		tokenizer = createTokenizer(input, NORMAL_PREFERENCE);
-		tokenizer.readColumns(columns);
-
-		assertTrue(columns.size() == 4);
-		assertEquals("one", columns.get(0));
-		assertEquals("two", columns.get(1));
-		assertEquals("field with \" quote char", columns.get(2));
-		assertEquals("four", columns.get(3));
-		assertEquals(input, tokenizer.getUntokenizedRow());
-	}
-
-	/**
-	 * Tests the readColumns() method with a quote character escaped with a backslash (non-RFC
-	 * style)
-	 */
-	@Test
-	public void testQuotedQuoteCharWithBackslashEscape() throws Exception {
-
-		final CsvPreference csvPref = new CsvPreference.Builder(NORMAL_PREFERENCE)
-				.setQuoteEscapeChar('\\')
-				.build();
-
-		final String input = "one,two,\"field with \\\" quote char\",four";
-		tokenizer = createTokenizer(input, csvPref);
-		tokenizer.readColumns(columns);
-
-		assertTrue(columns.size() == 4);
-		assertEquals("one", columns.get(0));
-		assertEquals("two", columns.get(1));
-		assertEquals("field with \" quote char", columns.get(2));
-		assertEquals("four", columns.get(3));
-		assertEquals(input, tokenizer.getUntokenizedRow());
-	}
-
-	/**
-	 * Tests the readColumns() method with an escape character following another escape char.
-	 */
-	@Test
-	public void testEscapedEscapeChar() throws Exception {
-
-		final CsvPreference csvPref = new CsvPreference.Builder(NORMAL_PREFERENCE)
-				.setQuoteEscapeChar('\\')
-				.build();
-
-		final String input = "\"field with \\\\ escape char\"";
-		tokenizer = createTokenizer(input, csvPref);
-		tokenizer.readColumns(columns);
-
-		assertTrue(columns.size() == 1);
-		assertEquals("field with \\ escape char", columns.get(0));
-		assertEquals(input, tokenizer.getUntokenizedRow());
-	}
-
-	/**
-	 * Tests the readColumns() method with an even number of escape characters
-	 * char.
-	 */
-	@Test
-	public void testEvenSeriesOfEscapeChars() throws Exception {
-
-		final CsvPreference csvPref = new CsvPreference.Builder(NORMAL_PREFERENCE)
-				.setQuoteEscapeChar('\\')
-				.build();
-
-		final String input = "\"\\\\\\\\\\\\\"";
-		tokenizer = createTokenizer(input, csvPref);
-		tokenizer.readColumns(columns);
-
-		assertTrue(columns.size() == 1);
-		assertEquals("\\\\\\", columns.get(0));
-		assertEquals(input, tokenizer.getUntokenizedRow());
-	}
-
-	/**
-	 * Tests the readColumns() method with an odd number of escape characters
-	 * char.
-	 */
-	@Test(expected = SuperCsvException.class)
-	public void testOddSeriesOfEscapeChars() throws Exception {
-
-		final CsvPreference csvPref = new CsvPreference.Builder(NORMAL_PREFERENCE)
-				.setQuoteEscapeChar('\\')
-				.build();
-
-		final String input = "\"\\\\\\\\\\\"";
-		tokenizer = createTokenizer(input, csvPref);
-		tokenizer.readColumns(columns);
-	}
-
-	/**
-	 * Tests the readColumns() method with a doubled escape char in an unquoted field, which
-	 * should show as two literal escape chars.
-	 *
-	 * Doubled escape chars only evaluate to a single escape char when they are in a quoted field.
-	 */
-	@Test
-	public void testDoubledEscapeCharInUnquotedFieldIsLiteral() throws Exception {
-		final CsvPreference csvPref = new CsvPreference.Builder(NORMAL_PREFERENCE)
-				.setQuoteEscapeChar('\\')
-				.build();
-
-		final String input = "\\\\";
-		tokenizer = createTokenizer(input, csvPref);
-		tokenizer.readColumns(columns);
-
-		assertTrue(columns.size() == 1);
-		assertEquals("\\\\", columns.get(0));
-		assertEquals(input, tokenizer.getUntokenizedRow());
-	}
-
-
-	/**
-	 * Tests the readColumns() method with a lone escape char outside of quoted fields
-	 */
-	@Test
-	public void testLoneEscapeCharInUnquotedField() throws Exception {
-		final CsvPreference csvPref = new CsvPreference.Builder(NORMAL_PREFERENCE)
-				.setQuoteEscapeChar('\\')
-				.build();
-
-		final String input = "one\\,\\two,thr\\ee";
-		tokenizer = createTokenizer(input, csvPref);
-		tokenizer.readColumns(columns);
-
-		assertTrue(columns.size() == 3);
-		assertEquals("one\\", columns.get(0));
-		assertEquals("\\two", columns.get(1));
-		assertEquals("thr\\ee", columns.get(2));
-		assertEquals(input, tokenizer.getUntokenizedRow());
-	}
-
-	/**
-	 * Tests the readColumns() method with a doubled escape char outside of quoted fields
-	 */
-	@Test
-	public void testDoubledEscapeCharInUnquotedField() throws Exception {
-		final CsvPreference csvPref = new CsvPreference.Builder(NORMAL_PREFERENCE)
-				.setQuoteEscapeChar('\\')
-				.build();
-
-		final String input = "one\\\\,\\\\two,thr\\\\ee";
-		tokenizer = createTokenizer(input, csvPref);
-		tokenizer.readColumns(columns);
-
-		assertTrue(columns.size() == 3);
-		assertEquals("one\\\\", columns.get(0));
-		assertEquals("\\\\two", columns.get(1));
-		assertEquals("thr\\\\ee", columns.get(2));
-		assertEquals(input, tokenizer.getUntokenizedRow());
-	}
-
-	/**
-	 * Tests the readColumns() method with an escape character preceding neither another escape
-	 * char nor a quote char.  In this situation, just pass the data through rather than
-	 * attempting to interpret the quote char.
-	 */
-	@Test
-	public void testEscapedNonEscapeChar() throws Exception {
-
-		final CsvPreference csvPref = new CsvPreference.Builder(NORMAL_PREFERENCE)
-				.setQuoteEscapeChar('\\')
-				.build();
-
-		final String input = "\"field with \\an escape char on neither escape nor quote\"";
-		tokenizer = createTokenizer(input, csvPref);
-		tokenizer.readColumns(columns);
-
-		assertTrue(columns.size() == 1);
-		assertEquals("field with \\an escape char on neither escape nor quote", columns.get(0));
-		assertEquals(input, tokenizer.getUntokenizedRow());
-	}
-
-	/**
-	 * Test double-quote char when in backslash-escape mode should throw exception
-	 */
-	@Test(expected = SuperCsvException.class)
-	public void testDoubleQuoteBackslashEscapeChar() throws Exception {
-
-		final CsvPreference csvPref = new CsvPreference.Builder(NORMAL_PREFERENCE)
-				.setQuoteEscapeChar('\\')
-				.build();
-
-		final String input = "\"field with an escaped quote \\\" and a \"\" double quote\"";
-		tokenizer = createTokenizer(input, csvPref);
-		tokenizer.readColumns(columns);
-
-		Assert.fail();
 	}
 	
 	/**
